requirements.txt
# Requirements for TimeTracker on Render
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.6
Werkzeug==3.1.3
SQLAlchemy==2.0.40
alembic==1.16.2
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==36.0.2
dnspython==2.7.0
fpdf==1.7.2
greenlet==3.2.1
itsdangerous==2.2.0
Mako==1.3.10
MarkupSafe==3.0.2
openpyxl==3.1.5
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.1
typing_extensions==4.15.0
et-xmlfile==2.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
APScheduler==3.10.4
eventlet==0.39.1
Flask-SocketIO==5.5.1
Flask-Mail==0.10.0
python-socketio==5.14.3
python-engineio==4.12.3
simple-websocket==1.1.0
bidict==0.23.1
supabase==2.23.0
python-magic==0.4.27
requests==2.32.5
